feat: moved things more into their own folders and updated the import paths
