+++ src/components/Header/partials/ButtonUploadJson/ButtonUploadJson.jsx
@@ -1,4 +1,4 @@
-import { SECURITY_LIMITS } from '../../../config/config.jsx';
+import { SECURITY_LIMITS } from '../../../../config/config.jsx';
 import './ButtonUploadJson.css';
 
 export function ButtonUploadJson({ onUploadJson }) {
