+++ src/components/RelationshipGraph/partials/RelationshipSVG/partials/SvgCatNodes/SvgCatNodes.jsx
@@ -3,7 +3,7 @@ import {
 	SEX_COLOR,
 	SEX_BG,
 	SEX_BG_HOVER,
-} from '../../../../../config/config.jsx';
+} from '../../../../../../config/config.jsx';
 import './SvgCatNodes.css';
 
 export default function SvgCatNodes({
